move ColorInput css into ColorInput.module.css
