+++ src/App.css
@@ -52,25 +52,6 @@
 	flex-wrap: wrap;
 }
 
-/*Color Input Component*/
-.colorInputContainer {
-	position: relative;
-}
-.inputContainer {
-	display: flex;
-	flex-direction: column;
-	flex-grow: 1;
-}
-.colorSwatch {
-	width: 2rem;
-	height: 2rem;
-	border-radius: 50%;
-	box-shadow: 2px 2px 5px var(--grey);
-	position: absolute;
-	left: 128px;
-	top: 1.75rem;
-	z-index: 1;
-}
 /* Delete button */
 .deleteButton {
 	position: absolute;
