+++ src/App.jsx
@@ -4,7 +4,7 @@ import { getRelativeLuminance, deepCopyArrayOfObject } from "./utils"
 import "./App.css"
 
 import { AiOutlineReload } from "react-icons/ai"
-import ColorInput from "./components/ColorInput"
+import ColorInput from "./components/ColorInput/ColorInput"
 import ContrastInfoCard from "./components/ContrastInfoCard"
 import Footer from "./components/Footer"
