+++ src/components/ColorInput/ColorInput.module.css
@@ -1,0 +1,19 @@
+/*Color Input Component*/
+.colorInputContainer {
+	position: relative;
+}
+.inputContainer {
+	display: flex;
+	flex-direction: column;
+	flex-grow: 1;
+}
+.colorSwatch {
+	width: 2rem;
+	height: 2rem;
+	border-radius: 50%;
+	box-shadow: 2px 2px 5px var(--grey);
+	position: absolute;
+	left: 128px;
+	top: 1.75rem;
+	z-index: 1;
+}
